overc: global fixups for oe-core master
This is a global fixup to build NOT boot, against oe-core master. Among
the changes are:

  - more override syntax fixups
  - updated bbappend versions
  - new bbappands for missing pkgconfig
  - creation of a bind-dhclient headers only package, so we can
    remove the bind recipe to update to the latest
  - drop old dependencies
  - make more changes overc distro dependent
  - remove ostree in favour of upstream recipe

+++ meta-cube/recipes-extended/shadow/shadow_%.bbappend
@@ -5,16 +5,18 @@
 # they may not be 100% portable to yocto.
 #
 
-do_install:append() {
+do_install:append:class-target() {
 	ln ${D}${sbindir}/useradd ${D}${sbindir}/adduser
 	ln ${D}${sbindir}/userdel ${D}${sbindir}/deluser
 
 	ln ${D}${sbindir}/groupadd ${D}${sbindir}/addgroup
 	ln ${D}${sbindir}/groupdel ${D}${sbindir}/delgroup
 
-	cd ${D}${mandir}/man8 && ln -s useradd.8 adduser.8
-	cd ${D}${mandir}/man8 && ln -s userdel.8 deluser.8
+	if [ -d "${D}${mandir}/man8" ] ; then
+		cd ${D}${mandir}/man8 && ln -s useradd.8 adduser.8
+		cd ${D}${mandir}/man8 && ln -s userdel.8 deluser.8
 
-	cd ${D}${mandir}/man8 && ln -s groupadd.8 addgroup.8
-	cd ${D}${mandir}/man8 && ln -s groupdel.8 delgroup.8
+		cd ${D}${mandir}/man8 && ln -s groupadd.8 addgroup.8
+		cd ${D}${mandir}/man8 && ln -s groupdel.8 delgroup.8
+	fi
 }
